# Conflicts:
#	arrows/ocv/detect_features_FAST.cxx
#	arrows/ocv/detect_features_FAST.h
#	arrows/ocv/track_features_klt.cxx
#	sprokit/processes/core/close_loops_process.cxx
#	sprokit/processes/core/detect_features_if_keyframe_process.cxx
#	sprokit/processes/core/keyframe_selection_process.cxx
#	vital/types/feature_track_set.cxx
#	vital/types/track.cxx

+++ arrows/core/keyframe_selector_basic.cxx
@@ -62,7 +62,7 @@ public:
     }
     if (config->has_value("keyframe_min_feature_count"))
     {
-      keyframe_min_feature_count = config->get_value<int>(
+      keyframe_min_feature_count = config->get_value<size_t>(
         "keyframe_min_feature_count");
     }
   }
@@ -118,7 +118,7 @@ public:
   bool a_keyframe_was_selected(
     kwiver::vital::track_set_sptr tracks);
 
-  int keyframe_min_feature_count;
+  size_t keyframe_min_feature_count;
   float fraction_tracks_lost_to_necessitate_new_keyframe;
 
   kwiver::vital::logger_handle_t m_logger;
